pyproject.toml
[tool.poetry]
name = "zdppy_yaml"
version = "0.1.0"
description = ""
authors = ["zhangdapeng <pygosuperman@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
